Bug #33680475 Missing checks in test wrapper can cause an invalid access exception
Protect against invalid access to class members in TearDown methods, because they can end-up not being initialized on SetUp (if for example, a previous error prevented the shell from connecting).

+++ unittest/modules/adminapi/common/metadata_management_t.cc
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2019, 2021, Oracle and/or its affiliates.
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License, version 2.0,
@@ -60,10 +60,13 @@ class Metadata_management_test : public Shell_core_test_wrapper {
   }
 
   void TearDown() override {
-    m_session->execute("DROP SCHEMA IF EXISTS mysql_innodb_cluster_metadata");
-    m_session->execute(
-        "DROP SCHEMA IF EXISTS mysql_innodb_cluster_metadata_bkp");
-    m_session->close();
+    if (m_session) {
+      m_session->execute("DROP SCHEMA IF EXISTS mysql_innodb_cluster_metadata");
+      m_session->execute(
+          "DROP SCHEMA IF EXISTS mysql_innodb_cluster_metadata_bkp");
+      m_session->close();
+    }
+
     Shell_core_test_wrapper::TearDown();
   }
